AbstractMem library to v1.4

+++ src/core/UNode.pas
@@ -801,7 +801,7 @@ function TNode.TryFindPublicSaleAccount(AMaximumPrice: Int64; APreventRaceCondit
   // Finds an account at or below argument purchase price (or returns false)
   // APreventRaceCondition: When True will return a random account in valid range price
   // Limitations: Account must be >0
-var LtempAccNumber : Integer;
+var LtempAccNumber : Int64;
   LLastValidAccount, LCurrAccount : TAccount;
   LContinueSearching : Boolean;
 begin
